Fix type issue

+++ src/loss/las_loss.py
@@ -8,4 +8,4 @@ class LASLossWrapper(torch.nn.Module):
     def forward(
         self, log_probs : torch.Tensor, text_encoded: torch.Tensor, **batch
     ) -> dict[str, torch.Tensor]:
-        return {"loss": torch.nn.CrossEntropyLoss()(log_probs.reshape(-1, log_probs.size(-1)), text_encoded.view(-1))}
+        return {"loss": torch.nn.CrossEntropyLoss()(log_probs.reshape(-1, log_probs.size(-1)), text_encoded.to(torch.long).view(-1))}
